+++ apps/block_scout_web/test/block_scout_web/controllers/api/v2/transaction_controller_test.exs
@@ -109,7 +109,7 @@ defmodule BlockScoutWeb.API.V2.TransactionControllerTest do
       |> with_block()
 
       auth = build(:auth)
-      address = insert(:address)
+      insert(:address)
       {:ok, user} = UserFromAuth.find_or_create(auth)
 
       conn = Plug.Test.init_test_session(conn, current_user: user)
@@ -234,15 +234,14 @@ defmodule BlockScoutWeb.API.V2.TransactionControllerTest do
         |> insert()
         |> with_block()
 
-      tt =
-        insert(:token_transfer,
-          transaction: tx,
-          block: tx.block,
-          block_number: tx.block_number,
-          token_contract_address: token.contract_address,
-          token_ids: Enum.map(0..50, fn x -> x end),
-          amounts: Enum.map(0..50, fn x -> x end)
-        )
+      insert(:token_transfer,
+        transaction: tx,
+        block: tx.block,
+        block_number: tx.block_number,
+        token_contract_address: token.contract_address,
+        token_ids: Enum.map(0..50, fn x -> x end),
+        amounts: Enum.map(0..50, fn x -> x end)
+      )
 
       request = get(conn, "/api/v2/transactions/" <> to_string(tx.hash))
 
@@ -991,6 +990,20 @@ defmodule BlockScoutWeb.API.V2.TransactionControllerTest do
     assert to_string(token_transfer.block_hash) == json["block_hash"]
     assert to_string(token_transfer.log_index) == json["log_index"]
     assert check_total(Repo.preload(token_transfer, [{:token, :contract_address}]).token, json["total"], token_transfer)
+  end
+
+  defp compare_item(%Transaction{} = transaction, json, wl_names) do
+    assert to_string(transaction.hash) == json["hash"]
+    assert transaction.block_number == json["block"]
+    assert to_string(transaction.value.value) == json["value"]
+    assert Address.checksum(transaction.from_address_hash) == json["from"]["hash"]
+    assert Address.checksum(transaction.to_address_hash) == json["to"]["hash"]
+
+    assert json["to"]["watchlist_names"] ==
+             if(wl_names[transaction.to_address_hash], do: [wl_names[transaction.to_address_hash]], else: [])
+
+    assert json["from"]["watchlist_names"] ==
+             if(wl_names[transaction.from_address_hash], do: [wl_names[transaction.from_address_hash]], else: [])
   end
 
   defp check_paginated_response(first_page_resp, second_page_resp, txs) do